合并dc版本

+++ Aps-Api/cx-api/src/main/java/com/zlt/aps/cx/api/domain/dto/BomInfoDto.java
@@ -104,6 +104,6 @@ public class BomInfoDto extends ApsBaseDto {
 
     /** 删除标识 */
     @ApiModelProperty(value = "删除标识")
-    private Long isDelete;
+    private Integer isDelete;
 
 }
